# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bizosto ERP API
  version: 1.0.0
  description: >-
    Multi-tenant API contract generated from App Router route handlers.
servers:
  - url: /
security:
  - BearerAuth: []
...
paths:
...
  /api/admin/finance/invoices/create:
    post:
      summary: POST /api/admin/finance/invoices/create
      tags:
        - Core
      responses:
        '200':
          description: Request completed successfully.
        '401':
          description: Unauthorized or invalid session.
        '403':
          description: Forbidden for current tenant or role.
...
  /api/admin/projects/create:
    post:
      summary: POST /api/admin/projects/create
      tags:
        - Core
      responses:
        '200':
          description: Request completed successfully.
        '401':
          description: Unauthorized or invalid session.
        '403':
          description: Forbidden for current tenant or role.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT